requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
pipenv==2022.5.2
platformdirs==2.5.2
requests==2.27.1
six==1.16.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.1.2
dash
matplotlib
plotly_express ==0.4.0
pandas
tweepy
textblob
